Implementation tweaks and simplification for release
