+++ .eslintrc.json
@@ -16,6 +16,7 @@
   },
   "rules": {
     "arrow-parens": 0,
+    "global-require": 0,
     "no-plusplus": 0,
     "no-unused-vars": [
       "error",
