+++ package.json
@@ -60,6 +60,7 @@
     "rimraf": "^2.5.4"
   },
   "dependencies": {
+    "core-js": "^2.4.1",
     "lodash.isequal": "^4.1.4",
     "object-assign": "^4.1.0",
     "util-inspect": "^0.1.8"
