+++ src/utils/Map.js
@@ -1,0 +1,5 @@
+export default (
+  typeof Map !== 'undefined'
+    ? Map
+    : require('core-js/library/es6/map')
+);
